feat: Adds ability to append data or clean table first

+++ airflow/plugins/operators/load_dimension.py
@@ -19,6 +19,7 @@ class LoadDimensionOperator(BaseOperator):
                  fields='',
                  redshift_conn_id='',
                  load_dimension='',
+                 append_data = False,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
@@ -27,13 +28,17 @@ class LoadDimensionOperator(BaseOperator):
         self.fields = fields
         self.redshift_conn_id = redshift_conn_id
         self.load_dimension = load_dimension
+        self.append_data = append_data
 
     def execute(self, context):
         self.log.info(f'Loading {self.table} dimensions in redshift')
         redshift = PostgresHook(postgres_conn_id = self.redshift_conn_id)
 
-        self.log.info('clearing data from Redshift table for new data')
-        redshift.run("DELETE FROM {}".format(self.table))
+
+
+        if self.append_data == False:
+            self.log.info('clearing data from Redshift table for new data')
+            redshift.run("DELETE FROM {}".format(self.table))
 
         dimensions_table_insert = LoadDimensionOperator.dimensions_table_insert.format(
             destination_table = self.table,
