+++ airflow/plugins/operators/load_fact.py
@@ -18,6 +18,7 @@ class LoadFactOperator(BaseOperator):
                  fields = '',
                  redshift_conn_id = '',
                  load_facts_sql = '',
+                 append_data = False,
                  *args, **kwargs):
 
         super(LoadFactOperator, self).__init__(*args, **kwargs)
@@ -26,13 +27,15 @@ class LoadFactOperator(BaseOperator):
         self.fields = fields
         self.redshift_conn_id = redshift_conn_id
         self.load_facts_sql = load_facts_sql
+        self.append_data = append_data
 
     def execute(self, context):
         self.log.info('Loading facts data into redshift facts_table')
         redshift = PostgresHook(postgres_conn_id = self.redshift_conn_id)
 
-        self.log.info('clearing data from Redshift table for new data')
-        redshift.run("DELETE FROM {}".format(self.table))
+        if self.append_data == False:
+            self.log.info('clearing data from Redshift table for new data')
+            redshift.run("DELETE FROM {}".format(self.table))
 
         facts_table_insert = LoadFactOperator.facts_table_insert.format(
             destination_facts_table = self.table,
